export types (#69)
