+++ src/types/syllabus.ts
@@ -1,0 +1,12 @@
+export type Syllabus = {
+  lecture_title: string;
+  year: string;
+  credit: string;
+  term: string;
+  person: string;
+  numbering: string;
+  department: string;
+  url: string;
+  dow: string;
+  period: string;
+};
